% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_wrapper.R
\name{LRMultiClass}
\alias{LRMultiClass}
\title{Multi logistic regression with ridge penalty}
\usage{
LRMultiClass(X, y, numIter = 50, eta = 0.1, lambda = 1, beta_init = NULL)
}
\arguments{
\item{X}{mat[n, p]: n row of data, p col of covariates}

\item{y}{vec[n]: each element is lie on 0 ~ K-1, denoting the class of the sample}

\item{numIter}{number of iterations, default 50}

\item{eta}{damping parameter, default 0.1, must be positive}

\item{lambda}{ridge parameter, default 1, must be non negative}

\item{beta_init}{mat[p, k]: starting beta values, if null, initialize with 0 matrix.}
}
\value{
out - list of updated beta and objective after iterations
}
\description{
Multi logistic regression with ridge penalty
}
\examples{
n1 = 50
p1 = 30
k1 = 10
X1 = matrix(rnorm(n1, (p1 - 1)), n1, p1 - 1)
X1 = cbind(rep(1, n1), X1)
y1 = sample(k1, n1, replace = TRUE) - 1
LRMultiClass(X1, y1, 50, 0.1, 1, NULL)

beta_init = matrix(1, p1, k1)
LRMultiClass(X1, y1, 100, 0.5, 0.5, beta_init)
}
