% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_wrapper.R
\name{MyKmeans}
\alias{MyKmeans}
\title{K-means algorithm}
\usage{
MyKmeans(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{mat[n, p]; n row of data with p covariate}

\item{K}{int; number of clusters}

\item{M}{mat[K, p]; cluster centers with default NULL; If NULL, initialize K data randomly from X}

\item{numIter}{int; number of iteration with default 100}
}
\value{
Y vec[n]; Y[i] is assigned cluster number for ith data
}
\description{
K-means algorithm
}
\examples{
n1 = 40
p1 = 20
K1 = 5
X1 = matrix(rnorm(n1 * p1), n1, p1)
MyKmeans(X1, K1, NULL, 100)

n2 = 50
p2 = 30
K2 = 10
X2 = matrix(rnorm(n2 * p2), n2, p2)
MyKmeans(X2, K2, NULL, 50)
}
